requirements.txt
streamlit==1.35.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.26.4
pickle-mixin==1.0.2
python-dotenv==1.0.1
langchain==0.1.16
langchain-google-genai==0.0.10
